requirements.txt
faker 
loguru
mysql-connector-python
pydantic
pandas
pymongo
psycopg2-binary


